requirements.txt
langchain==0.2.6
langchain-core==0.2.11
langchain-groq==0.1.6
python-dotenv==1.0.1
streamlit==1.36.0